Unified all model optimisers to just one
